+++ src/frontend/magnify/src/components/profile/index.ts
@@ -1,4 +1,5 @@
 export { default as AvatarForm, AvatarFormProps } from './AvatarForm';
+export { default as DeleteAccountBlock } from './DeleteAccountBlock';
 export { default as IdentityBlock, IdentityBlockProps } from './IdentityBlock';
 export { default as IdentityForm, IdentityFormProps } from './IdentityForm';
 export { default as PasswordUpdateBlock } from './PasswordUpdateBlock';